% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_information.R
\name{has_colnames}
\alias{has_colnames}
\title{Check if object has colnames}
\usage{
has_colnames(x)
}
\arguments{
\item{x}{object.}
}
\value{
TRUE or FALSE
}
\description{
Check if object has colnames. Note it is not checked first whether the object is a matrix or dataframe.
}

