% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_algebra.R
\name{has_intercept}
\alias{has_intercept}
\title{Check if model matrix has intercept}
\usage{
has_intercept(x)
}
\arguments{
\item{x}{value.}
}
\value{
TRUE or FALSE
}
\description{
Checks whether the value is a full rank matrix with an intercept term as the first colum.
}

