% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_information.R
\name{has_rownames}
\alias{has_rownames}
\title{Check if object has rownames}
\usage{
has_rownames(x)
}
\arguments{
\item{x}{object.}
}
\value{
TRUE or FALSE
}
\description{
Check if object has rownames. Note it is not checked first whether the object is a matrix or dataframe.
}

