% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_arguments.R
\name{in_option}
\alias{in_option}
\title{Check if value is a valid option}
\usage{
in_option(x, options)
}
\arguments{
\item{x}{value.}

\item{options}{possible options for x to take.}
}
\value{
TRUE or FALSE
}
\description{
Checks whether the value is scalar and in the supplied options. No class checks are performed.
}

