% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_arguments.R
\name{is_class}
\alias{is_class}
\title{Check if object is valid class}
\usage{
is_class(x, options)
}
\arguments{
\item{x}{value.}

\item{options}{possible classes for x.}
}
\value{
TRUE or FALSE
}
\description{
Checks whether the object is valid class.
}

