% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_algebra.R
\name{is_fullrank}
\alias{is_fullrank}
\title{Check if model matrix is full rank}
\usage{
is_fullrank(x)
}
\arguments{
\item{x}{value.}
}
\value{
TRUE or FALSE
}
\description{
Checks whether the value is full rank matrix.
}

